Repostiory: Fix bugs after move package class into its own file
